New scene: Prototype level is complete with: -Graveyard -Orc Camp -City -Dwarf Camp -Wandering mobs
Notes:
Each camp is a separate group and each group of enemies are a subgroup so it'll be easy to find them.
It's 100x100
Trees and Vegetation are not great, but will do.
Collision is a little funky and will need to do a quality pass at some point so you don't get stuck on shit as much.